refactor(H3Data): Refactors code related H3 Maps
Code for serving H3 data maps has been refactored into a new service/repository, plus smaller refactors
